Backend/reactbot/requirements.txt
Django
djangorestframework
markdown
django-filter
django-cors-headers
requests
djangorestframework-simplejwt
django-countries
django-extensions
pyparsing
pydot